refactor(projectTree): extract search index into ProjectSearchIndex.ts
Moves the self-contained search subsystem (build/rank/search + LRU cache + the 4
free helpers cacheKeyForSearchConfig/partitionExcludePatterns/expandDirPatterns/
pathHitsIgnoredName) off ProjectTreeService into a new ProjectSearchIndex class,
threaded with its deps (WslBridgeClient, resolveEntryPath callback). projectTree.ts
drops 1106->870; search/rank logic now has an isolatable home with its own cache
so it can be unit-tested (previously zero coverage). Pure structural move —
ranking, cache keying, LRU eviction, and invalidation prefix-matching byte-identical.
projectTree.test.ts (18) green.
